Handle all the possible intersection types
There is now a recursive function extracting the points from
intersections, in an exhaustive way (excluding polygons).

The crossing restriction has been removed because it is wrong
for overlapping segments: the corresponding test now passes.
A notable consequence is the change in handling endpoints:
intersections with endpoints are now accepted.

Issue: tomalrussel/snkit#15